Plumb user driven updates to properties through a stream of updates.
The simple goal: allow multiple updates to a set of properties in a single
task to culminate into a single update being sent to FlightGear. This is
done in a microtask and syncrhonus streams to get'er done with minimal
delay. Be warnned: there is no gating here because the Property tree
doesn't care / know about your flightgear settings. Implementation of
gating is left to the user.

This also moves Property parsing / updates to the class instead of
coupling that to the property tree.