Widgets...
Ruminating about simple widgets. For radios, there's the
RadioFrequencyWidget and the RadilWidget - useful for making a simple
NAV page with active & standby frequencies (of which you can edit the
standby and then swap the two) and set the radial for positing. For
setting the altitude, its a simpler version of the RFW - just modify a
value by 100s or 1000s. Each of these widgets lets you visually see when
they are selected by highlighting the 10s poistion knobs can adjust.

Then there's a poorly named SwitchWidget.. which is really a button...
which can be updated.  it shows highllight by adding [] around the
text... so "foo" is "[foo]" when focused.. followed by color filling
when actually selected.