Remember properties on each page; fix radial.adjust()
Alright, this works.. but it sucks. Why? because I'm getting a lot of
noise from FlightGear which causes the values to bounce back and forth.
This is less than ideal as knob turns are bounced. To fix this, I'll
need to ignore incoming property updates for knobs - but leave switches
alone.